vanna-streamlit-chatbot/requirements.txt
vanna
streamlit
python-dotenv
